+++ learning/tour-of-beam/backend/internal/storage/adapter.go
@@ -22,20 +22,6 @@ import (
 	"cloud.google.com/go/datastore"
 )
 
-func sdkToKey(sdk tob.Sdk) string {
-	switch sdk {
-	case tob.SDK_GO:
-		return "SDK_GO"
-	case tob.SDK_PYTHON:
-		return "SDK_PYTHON"
-	case tob.SDK_JAVA:
-		return "SDK_JAVA"
-	case tob.SDK_SCIO:
-		return "SDK_SCIO"
-	}
-	panic(fmt.Sprintf("Undefined key for sdk: %s", sdk))
-}
-
 func pgNameKey(kind, nameId string, parentKey *datastore.Key) (key *datastore.Key) {
 	key = datastore.NameKey(kind, nameId, parentKey)
 	key.Namespace = PgNamespace
@@ -45,7 +31,7 @@ func pgNameKey(kind, nameId string, parentKey *datastore.Key) (key *datastore.Ke
 // Get entity key from sdk & entity ID
 // SDK_JAVA_{entityID}.
 func datastoreKey(kind string, sdk tob.Sdk, id string, parent *datastore.Key) *datastore.Key {
-	name := fmt.Sprintf("%s_%s", sdkToKey(sdk), id)
+	name := fmt.Sprintf("%s_%s", sdk.StorageID(), id)
 	return pgNameKey(kind, name, parent)
 }
 
@@ -78,9 +64,7 @@ func MakeGroupNode(group *tob.Group, order, level int) *TbLearningNode {
 		return nil
 	}
 	return &TbLearningNode{
-		// ID doesn't make much sense for groups,
-		// but we have to define it to include in queries
-		Id:    group.Title,
+		Id:    group.Id,
 		Title: group.Title,
 
 		Type:  tob.NODE_GROUP,
@@ -110,12 +94,13 @@ func FromDatastoreUnit(tbUnit *TbLearningUnit, id, title string) *tob.Unit {
 }
 
 // Depending on the projection, we either convert TbLearningGroup to a model
-// Or we use common field Title to make it.
-func FromDatastoreGroup(tbGroup *TbLearningGroup, title string) *tob.Group {
+// Or we use common fields Id, Title to make it.
+func FromDatastoreGroup(tbGroup *TbLearningGroup, id, title string) *tob.Group {
 	if tbGroup == nil {
-		return &tob.Group{Title: title}
+		return &tob.Group{Id: id, Title: title}
 	}
 	return &tob.Group{
+		Id:    tbGroup.Id,
 		Title: tbGroup.Title,
 	}
 }
@@ -126,7 +111,7 @@ func FromDatastoreNode(tbNode TbLearningNode) tob.Node {
 	}
 	switch tbNode.Type {
 	case tob.NODE_GROUP:
-		node.Group = FromDatastoreGroup(tbNode.Group, tbNode.Title)
+		node.Group = FromDatastoreGroup(tbNode.Group, tbNode.Id, tbNode.Title)
 	case tob.NODE_UNIT:
 		node.Unit = FromDatastoreUnit(tbNode.Unit, tbNode.Id, tbNode.Title)
 	default: