+++ runners/google-cloud-dataflow-java/worker/src/main/java/org/apache/beam/runners/dataflow/worker/util/JfrInterop.java
@@ -25,6 +25,7 @@ import java.time.Instant;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
+import org.apache.beam.vendor.guava.v26_0_jre.com.google.common.util.concurrent.ThreadFactoryBuilder;
 
 /**
  * Exposes methods to interop with JFR. This is only supported on java 9 and up, java 8 does not
@@ -35,7 +36,9 @@ import java.util.concurrent.Executors;
  */
 class JfrInterop {
   // ensure only a single JFR profile is running at once
-  private static final ExecutorService JFR_EXECUTOR = Executors.newSingleThreadExecutor();
+  private static final ExecutorService JFR_EXECUTOR =
+      Executors.newSingleThreadExecutor(
+          new ThreadFactoryBuilder().setDaemon(true).setNameFormat("JFRprofile-thread").build());
 
   private final Constructor<?> recordingCtor;
   private final Method recordingStart;