+++ sdks/python/apache_beam/typehints/typehints.py
@@ -287,8 +287,8 @@ class SequenceTypeConstraint(IndexableTypeConstraint):
             'instead received an instance of type %s.' % (
                 repr(self),
                 index,
-                _unified_repr(self._sequence_type),
-                _unified_repr(self.inner_type),
+                repr(self._sequence_type),
+                repr(self.inner_type),
                 elem.__class__.__name__))
       except CompositeTypeHintError as e:
         raise CompositeTypeHintError(
@@ -392,27 +392,6 @@ def validate_composite_type_param(type_param, error_msg_prefix):
         '%s must be a non-sequence, a type, or a TypeConstraint. %s'
         ' is an instance of %s.' %
         (error_msg_prefix, type_param, type_param.__class__.__name__))
-
-
-# TODO(https://github.com/apache/beam/issues/20982): Remove this function and
-# use plain repr() instead.
-def _unified_repr(o):
-  """Given an object return a qualified name for the object.
-
-  This function closely mirrors '__qualname__' which was introduced in
-  Python 3.3. It is used primarily to format types or object instances for
-  error messages.
-
-  Args:
-    o: An instance of a TypeConstraint or a type.
-
-  Returns:
-    A qualified name for the passed Python object fit for string formatting.
-  """
-  if isinstance(o, (TypeConstraint, type(None))) or not hasattr(o, '__name__'):
-    return repr(o)
-  else:
-    return o.__name__
 
 
 def check_constraint(type_constraint, object_instance):
@@ -533,7 +512,7 @@ class UnionHint(CompositeTypeHint):
     def __repr__(self):
       # Sorting the type name strings simplifies unit tests.
       return 'Union[%s]' % (
-          ', '.join(sorted(_unified_repr(t) for t in self.union_types)))
+          ', '.join(sorted(repr(t) for t in self.union_types)))
 
     def inner_types(self):
       for t in self.union_types:
@@ -565,7 +544,7 @@ class UnionHint(CompositeTypeHint):
           '%s type-constraint violated. Expected an instance of one of: %s, '
           'received %s instead.%s' % (
               repr(self),
-              tuple(sorted(_unified_repr(t) for t in self.union_types)),
+              tuple(sorted(repr(t) for t in self.union_types)),
               instance.__class__.__name__,
               error_msg))
 
@@ -673,7 +652,7 @@ class TupleHint(CompositeTypeHint):
       super().__init__(type_param, tuple)
 
     def __repr__(self):
-      return 'Tuple[%s, ...]' % _unified_repr(self.inner_type)
+      return 'Tuple[%s, ...]' % repr(self.inner_type)
 
     def _consistent_with_check_(self, sub):
       if isinstance(sub, TupleConstraint):
@@ -696,8 +675,7 @@ class TupleHint(CompositeTypeHint):
       return hash(self.tuple_types)
 
     def __repr__(self):
-      return 'Tuple[%s]' % (
-          ', '.join(_unified_repr(t) for t in self.tuple_types))
+      return 'Tuple[%s]' % (', '.join(repr(t) for t in self.tuple_types))
 
     def _inner_types(self):
       for t in self.tuple_types:
@@ -737,11 +715,8 @@ class TupleHint(CompositeTypeHint):
           raise CompositeTypeHintError(
               '%s hint type-constraint violated. The type of element #%s in '
               'the passed tuple is incorrect. Expected an instance of '
-              'type %s, instead received an instance of type %s.' % (
-                  repr(self),
-                  type_pos,
-                  _unified_repr(expected),
-                  actual.__class__.__name__))
+              'type %s, instead received an instance of type %s.' %
+              (repr(self), type_pos, repr(expected), actual.__class__.__name__))
         except CompositeTypeHintError as e:
           raise CompositeTypeHintError(
               '%s hint type-constraint violated. The type of element #%s in '
@@ -805,7 +780,7 @@ class ListHint(CompositeTypeHint):
       super().__init__(list_type, list)
 
     def __repr__(self):
-      return 'List[%s]' % _unified_repr(self.inner_type)
+      return 'List[%s]' % repr(self.inner_type)
 
   def __getitem__(self, t):
     validate_composite_type_param(t, error_msg_prefix='Parameter to List hint')
@@ -864,8 +839,7 @@ class DictHint(CompositeTypeHint):
       self.value_type = normalize(value_type)
 
     def __repr__(self):
-      return 'Dict[%s, %s]' % (
-          _unified_repr(self.key_type), _unified_repr(self.value_type))
+      return 'Dict[%s, %s]' % (repr(self.key_type), repr(self.value_type))
 
     def __eq__(self, other):
       return (
@@ -896,7 +870,7 @@ class DictHint(CompositeTypeHint):
                 repr(self),
                 incorrect_type[:-1],
                 incorrect_type,
-                _unified_repr(hinted_type),
+                repr(hinted_type),
                 inner_error_message))
       else:
         raise CompositeTypeHintError(
@@ -905,7 +879,7 @@ class DictHint(CompositeTypeHint):
                 repr(self),
                 incorrect_type[:-1],
                 incorrect_type,
-                _unified_repr(hinted_type),
+                repr(hinted_type),
                 incorrect_instance,
                 incorrect_instance.__class__.__name__))
 
@@ -986,7 +960,7 @@ class SetHint(CompositeTypeHint):
       super().__init__(type_param, set)
 
     def __repr__(self):
-      return 'Set[%s]' % _unified_repr(self.inner_type)
+      return 'Set[%s]' % repr(self.inner_type)
 
   def __getitem__(self, type_param):
     validate_composite_type_param(
@@ -1012,7 +986,7 @@ class FrozenSetHint(CompositeTypeHint):
             self).__init__(type_param, frozenset)
 
     def __repr__(self):
-      return 'FrozenSet[%s]' % _unified_repr(self.inner_type)
+      return 'FrozenSet[%s]' % repr(self.inner_type)
 
   def __getitem__(self, type_param):
     validate_composite_type_param(
@@ -1036,7 +1010,7 @@ class IterableHint(CompositeTypeHint):
             self).__init__(iter_type, abc.Iterable)
 
     def __repr__(self):
-      return 'Iterable[%s]' % _unified_repr(self.inner_type)
+      return 'Iterable[%s]' % repr(self.inner_type)
 
     def _consistent_with_check_(self, sub):
       if isinstance(sub, SequenceTypeConstraint):
@@ -1077,7 +1051,7 @@ class IteratorHint(CompositeTypeHint):
       self.yielded_type = normalize(t)
 
     def __repr__(self):
-      return 'Iterator[%s]' % _unified_repr(self.yielded_type)
+      return 'Iterator[%s]' % repr(self.yielded_type)
 
     def __eq__(self, other):
       return (
@@ -1106,10 +1080,8 @@ class IteratorHint(CompositeTypeHint):
       except SimpleTypeHintError:
         raise CompositeTypeHintError(
             '%s hint type-constraint violated. Expected a iterator of type %s. '
-            'Instead received a iterator of type %s.' % (
-                repr(self),
-                _unified_repr(self.yielded_type),
-                instance.__class__.__name__))
+            'Instead received a iterator of type %s.' %
+            (repr(self), repr(self.yielded_type), instance.__class__.__name__))
 
   def __getitem__(self, type_param):
     validate_composite_type_param(
@@ -1164,7 +1136,7 @@ class WindowedTypeConstraint(TypeConstraint, metaclass=GetitemConstructor):
           'is incorrect. Expected an instance of type %s, '
           'instead received an instance of type %s.' % (
               repr(self),
-              _unified_repr(self.inner_type),
+              repr(self.inner_type),
               instance.value.__class__.__name__))